added stack tracing to the chart interface

+++ inc/tracer.h
@@ -30,7 +30,7 @@
 #ifndef RANGE_CHECK
 #define RANGE_CHECK(VAR, MIN, MAX, ERROR_CODE, ERROR_CODE_STR)            \
   do {                                                                    \
-    if (!(VAR >= MIN && VAR < MAX)) {                                     \
+    if (!((int)VAR >= (int)MIN && (int)VAR < (int)MAX)) {                                     \
       printf("error: %s@%s:%d => %s\n", __func__, __FILENAME__, __LINE__, \
              ERROR_CODE_STR[ERROR_CODE]);                                 \
       return ERROR_CODE;                                                  \
